% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{buildparent}
\alias{buildparent}
\title{Build large parent raster}
\usage{
buildparent(x, rr)
}
\arguments{
\item{x}{extent}

\item{rr}{resolution}
}
\value{
SpatRaster
}
\description{
Create a large sparse raster from an extent and resolution. This grid is never meant to be populated with data
but used as a specification for smaller child windows that are aligned exactly within this larger, high-resolution parent.
}
\details{
Currently the resolution is required to be square.
}
