% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{airport_footprint}
\alias{airport_footprint}
\title{Calculate emissions per flight distance and type}
\usage{
airport_footprint(departure, arrival, flightClass = "Unknown",
  output = "co2e")
}
\arguments{
\item{departure}{{character} IATA code for outbound destination}

\item{arrival}{{character} IATA code for inbound destination}

\item{flightClass}{{character} flight class category, one of ""Unknown",  "Economy",
"Economy+", "Business" or "First"}

\item{output}{{character} emissions metric of the output}
}
\value{
numeric
}
\description{
Calculate emissions per flight distance and type
}
\examples{
airport_footprint("LAX", "PUS")
airport_footprint("LAX", "PUS", "First")
airport_footprint("LAX", "PUS", "First", "ch4")
}
