% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_footprint.R
\name{airport_footprint}
\alias{airport_footprint}
\title{Calculate emissions per flight distance and type using IATA airport codes}
\usage{
airport_footprint(departure, arrival, flightClass = "Unknown",
  output = "co2e")
}
\arguments{
\item{departure}{{character} IATA code for outbound destination}

\item{arrival}{{character} IATA code for inbound destination}

\item{flightClass}{{character} flight class category, one of "Unknown" "Economy", "Economy+", "Business" or "First". If no argument is included, "Unknown" is the default.}

\item{output}{{character} emissions metric of the output. For metrics that include radiative forcing, one of
\itemize{
\item "co2e" (carbon dioxide equivalent with radiative forcing) - default
\item "co2" (carbon dioxide with radiative forcing)
\item "ch4" (methane with radiative forcing)
\item "n2o" (nitrous oxide with radiative forcing)
\item Metrics without radiative forcing: "co2e_norf", "co2_norf", "ch4_norf", or "n2o_norf".
}}
}
\value{
a single numeric value expressed in kilograms
}
\description{
Calculate emissions per flight distance and type using IATA airport codes
}
\details{
The carbon footprint estimates are derived from the Department for Environment, Food & Rural Affiars (UK) 2019 Greenhouse Gas Conversion Factors for Business Travel (air): https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2019
}
\examples{
airport_footprint("LAX", "LHR")
airport_footprint("LAX", "LHR", "First")
airport_footprint("LAX", "LHR", "First", "ch4")
airport_footprint("LAX", "LHR", output = "ch4")
}
