% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong_footprint.R
\name{latlong_footprint}
\alias{latlong_footprint}
\title{Calculate emissions per flight distance and type using latitude and longitude}
\usage{
latlong_footprint(
  departure_lat,
  departure_long,
  arrival_lat,
  arrival_long,
  flightClass = "Unknown",
  output = "co2e"
)
}
\arguments{
\item{departure_lat}{{numeric} latitude of outbound destination}

\item{departure_long}{{numeric} longitude of outbound destination}

\item{arrival_lat}{{numeric} latitude of inbound destination}

\item{arrival_long}{{numeric} longitude of inbound destination}

\item{flightClass}{{character} flight class category, one of "Unknown" "Economy", "Economy+", "Business" or "First". If no argument is included, "Unknown" is the default.}

\item{output}{{character} emissions metric of the output. For metrics that include radiative forcing, one of
\itemize{
\item "co2e" (carbon dioxide equivalent with radiative forcing) - default
\item "co2" (carbon dioxide with radiative forcing)
\item "ch4" (methane with radiative forcing)
\item "n2o" (nitrous oxide with radiative forcing)
\item Metrics without radiative forcing: "co2e_norf", "co2_norf", "ch4_norf", or "n2o_norf".
#'
}}
}
\value{
a single numeric value expressed in kilograms
}
\description{
Calculate emissions per flight distance and type using latitude and longitude
}
\details{
The carbon footprint estimates are derived from the Department for Environment, Food & Rural Affiars (UK) 2019 Greenhouse Gas Conversion Factors for Business Travel (air): https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2019
}
\examples{
latlong_footprint(34.052235, -118.243683, 35.179554, 129.075638)
latlong_footprint(34.052235, -118.243683, 35.179554, 129.075638, "First")
latlong_footprint(34.052235, -118.243683, 35.179554, 129.075638, "First", "ch4")
latlong_footprint(34.052235, -118.243683, 35.179554, 129.075638, output = "ch4")
data \%>\% rowwise() \%>\% mutate(emissions = from_lat, from_long, to_lat, to_long, "Economy")
}
